Fix for libboost1.58
Compilation fails with libboost1.58 and you can't install 1.54 on
Ubuntu 16.04. I copied the code from this pull request in the
bitcoin repo https://github.com/bitcoin/bitcoin/pull/6114/files.
In order to compile it you have to run configure with additional
parameters (I added them to the readme.md in the ubuntu16.04 branch)